Code Cleanup
this is a merely a cleanup of the code that divides and organises the functions and variables.
No actual modification to program logic has been made (not intentionally at least). However there has been a modification and now all headered "cout/cin/endl" functions have now got "std::" in front of it to make up for the lack of "using namespace std;" line in the headers.